fixed admin course image, optimizing course, optimizing query

+++ reviewer/views/course.py
@@ -2,10 +2,12 @@
 from ..models import Course, Comment, Likes, Announcement
 from ..forms import CourseForm, CommentForm
 
+from django.conf import settings
+from django.core.exceptions import PermissionDenied
 from django.http import Http404, JsonResponse
 from django.shortcuts import redirect, render
-
 from django.urls import reverse
+
 from math import ceil
 from PIL import Image
 
@@ -23,43 +25,37 @@ def course(request, csubj, cnum):
 
 def coursecpage(request, csubj, cnum, catchar = 'l', cpage = 1):
 
-	if (catchar == None):
+	if not catchar:
 		catchar = "l"
 		cpage = 1
 
-	if (cpage == ""):
-		cpage = 1
-
+	if not cpage: cpage = 1
 	cpage = int(cpage)
-	page_ct_limit = 10	
+	page_ct_limit = settings.IMPORT_COMMENT_CT	
 
 	coursefilter = Course.objects.filter(code__iexact=csubj).filter(number__iexact=str(cnum))
 
 	if (len(coursefilter) > 0):
-		if (cpage < 1):
-			cpage = 1
+		if cpage < 1: cpage = 1
 
 		startindex = (cpage - 1)*page_ct_limit
-		course_comments_filtered = coursefilter[0].comment_set.order_by('-date_posted')[startindex:startindex+page_ct_limit]
-		course_commentstotal = len(coursefilter[0].comment_set.order_by('-date_posted'))
+		all_courses = coursefilter[0].comment_set.order_by('-date_posted')
+		course_comments_filtered = all_courses[startindex:startindex+page_ct_limit]
+		course_commentstotal = len(all_courses)
 
 		page_ct = int(ceil(course_commentstotal/page_ct_limit))
 
-		if (cpage > page_ct and page_ct != 0):
+		if cpage > page_ct and page_ct != 0:
 			return redirect('course', csubj, cnum)
-
 		else:
 			if request.method == "POST":
 
 				data = request.POST
 				resultid = None 
 
-				response = redirect('/course/'+csubj+'/'+ str(cnum) + '/c/1', csubj, cnum)
-
-				commentform = CommentForm(data, request.FILES, request=request)
+				response = redirect( reverse('coursecpage', args=[csubj, str(cnum), 'c', '1']) )
 
 				if request.user.is_authenticated:
-					
 					image_uploaded = request.FILES.get('image', None)
 
 					try:
@@ -70,12 +66,11 @@ def coursecpage(request, csubj, cnum, catchar = 'l', cpage = 1):
 						image_uploaded = None
 
 					# Add input validation
-					new_comment = Comment(course_attr=coursefilter[0], user_attr=request.user, body=data['body'], image=image_uploaded)	
-					new_comment.save()
+					Comment.objects.create(course_attr=coursefilter[0], user_attr=request.user, body=data['body'], image=image_uploaded)	
 					resultid = new_comment.id
 
 					if cpage == 1:
-						data = {'commentid': resultid}
+						data = {'commentid': resultid }
 						if new_comment.image:
 							data['image'] =  new_comment.image.url
 						response = JsonResponse(data)
@@ -93,7 +88,6 @@ def coursecpage(request, csubj, cnum, catchar = 'l', cpage = 1):
 
 					try:
 						last_comment = coursefilter[0].comment_set.order_by('-date_posted')[startindex+page_ct_limit-1]
-
 						likedstat = last_comment.likes_set.filter(user_attr=request.user).exists()
 
 						if last_comment:
@@ -104,7 +98,7 @@ def coursecpage(request, csubj, cnum, catchar = 'l', cpage = 1):
 								'body' : str(last_comment.body),
 								'date' : str(last_comment.date_posted),
 								'liked' : likedstat,
-								'like_ct' : len(last_comment.likes_set.all())
+								'like_ct' : last_comment.likes_set.all().count()
 							 }
 
 							if last_comment.image:
@@ -118,25 +112,18 @@ def coursecpage(request, csubj, cnum, catchar = 'l', cpage = 1):
 
 			elif request.method == "GET":
 
-				liked_comments = []
-
+				liked_comments = None
 				commentform = CommentForm(auto_id="comment-form", request=request)
 
 				if request.user.is_authenticated:
-					
-
-					user_likes = Likes.objects.filter(user_attr=request.user)
-
-					for i in user_likes:
-						liked_comments.append(int(i.comment.id))
-
+					liked_comments = list(Likes.objects.filter(user_attr=request.user).values_list('comment__id', flat=True))
 
 				context = {
 						'course_filt': coursefilter[0],
 						'course_comment_count': course_commentstotal,
 						'course_comments': course_comments_filtered,
 						'comment_form': commentform,
-						'liked_comments': str(liked_comments),
+						'liked_comments': liked_comments,
 						'section': catchar,
 						'page_count' : page_ct,
 						'cpage' : cpage,
@@ -152,32 +139,24 @@ def comment_like(request, csubj, cnum):
 	if request.method == "POST":
 
 		data = request.POST
-		resultid = None 
-
-		response = redirect('/course/'+csubj+'/'+ str(cnum) + '/c/1', csubj, cnum)
+		resultid = None
 
 		if request.user.is_authenticated:
 
-			comment_findid = int(request.body.decode("utf-8").split("-")[1])
-			comm_like = Comment.objects.filter(pk=comment_findid).first()	
+			print(data)
+			comment_findid = data.get('commentID').split("-")[1]
+			comm_all_like = Comment.objects.filter(pk=comment_findid).first().likes_set.all()
+			like_state = comm_all_like.filter(user_attr=request.user).first()
 
-			like_state = Likes.objects.filter(comment=comm_like, user_attr=request.user)
-
-			if like_state.exists():
-				like_state.delete()
-
-			else:
-				new_like = Likes(comment=comm_like, user_attr=request.user)	
-				new_like.save()
-
-			like_state = not like_state.exists()
+			like_state.delete() if like_state else Likes.objects.create(comment=comm_all_like[0].comment, user_attr=request.user)	
+			like_state = not bool(like_state)
 
 			like_count_callback = {
-				'count': len(comm_like.likes_set.all()), 
+				'count': len(comm_all_like), 
 				'state': like_state
 			}
 
-			response = JsonResponse(like_count_callback)
-
-		return response
+			return JsonResponse(like_count_callback)
+		else:
+			raise PermissionDenied()
